Add separator script
This started out as a copy of the preview notebook.
We implemented some logic to detect 6 regions in the MIP viewed along
the rotation axis.
Then we started to separate the labels from the bottom of the scan.
Then labeling the detected regions, sorting them to the correct scan
name and extracting each region out of the stack of reconstructions.
This is then written to disk, first as a .zarr file for each region and
then subsequently as several stacks of PNGs for further work by Ben and
team.
The labeling at the bottom of the notebook has to be implemented now
correctly.