More intermediate output for faster results
We save the bottom MIP to disk so we don't have to calculate it over and
over.
This makes re-runs of the notebook faster.
We also read in the tracking XLS sheet from Ben, so we can start to
automatically doublecheck his and my naming of things to potentially
spot errors.